+++ django_icons/renderers/__init__.py
@@ -1,3 +1,4 @@
 from .base import BaseRenderer
 from .bootstrap3 import Bootstrap3Renderer
 from .fontawesome import FontAwesomeRenderer
+from .material import MaterialRenderer