+++ tests/test_fontawesome.py
@@ -13,6 +13,6 @@ class FontAwesomeTest(TestCase):
 
     def test_icons(self):
         self.assertEqual(
-            render_template('{% icon "user" size="lg" %}'),
+            render_template('{% icon "user" size="lg" renderer="FontAwesomeRenderer" %}'),
             '<i class="fa fa-user fa-lg"></i>',
         )